handle rtc error stream
